wip: reorg so there is an AppShell component
