+++ web/src/components/AppShell.tsx
@@ -1,0 +1,251 @@
+import * as React from "react";
+import { Link } from "@remix-run/react";
+import { Dialog, Transition } from "@headlessui/react";
+import {
+  Bars3BottomLeftIcon,
+  BeakerIcon,
+  BellIcon,
+  FolderIcon,
+  UsersIcon,
+  XMarkIcon,
+} from "@heroicons/react/24/outline";
+import { MagnifyingGlassIcon } from "@heroicons/react/20/solid";
+import clsx from "clsx";
+
+import hitsaveLogo from "~/images/hitsave_logo.svg";
+import { ProfileDropdown } from "~/components/ProfileDropdown";
+import { User } from "~/session.server";
+
+const { Fragment, useState } = React;
+
+interface AppShellProps {
+  user: User;
+  children?: React.ReactNode;
+}
+
+export const AppShell: React.FC<AppShellProps> = ({ user, children }) => {
+  const [sidebarOpen, setSidebarOpen] = useState(false);
+  const navigation = [
+    {
+      name: "Projects",
+      href: "/dashboard/projects",
+      icon: FolderIcon,
+      current: false,
+      comingSoon: true,
+    },
+    {
+      name: "Experiments",
+      href: "/dashboard/experiments",
+      icon: BeakerIcon,
+      current: false,
+    },
+    {
+      name: "Team",
+      href: "/dashboard/team",
+      icon: UsersIcon,
+      current: false,
+      comingSoon: true,
+    },
+  ];
+
+  return (
+    <>
+      <>
+        <div>
+          <Transition.Root show={sidebarOpen} as={Fragment}>
+            <Dialog
+              as="div"
+              className="relative z-40 md:hidden"
+              onClose={setSidebarOpen}
+            >
+              <Transition.Child
+                as={Fragment}
+                enter="transition-opacity ease-linear duration-300"
+                enterFrom="opacity-0"
+                enterTo="opacity-100"
+                leave="transition-opacity ease-linear duration-300"
+                leaveFrom="opacity-100"
+                leaveTo="opacity-0"
+              >
+                <div className="fixed inset-0 bg-gray-600 bg-opacity-75" />
+              </Transition.Child>
+
+              <div className="fixed inset-0 z-40 flex">
+                <Transition.Child
+                  as={Fragment}
+                  enter="transition ease-in-out duration-300 transform"
+                  enterFrom="-translate-x-full"
+                  enterTo="translate-x-0"
+                  leave="transition ease-in-out duration-300 transform"
+                  leaveFrom="translate-x-0"
+                  leaveTo="-translate-x-full"
+                >
+                  <Dialog.Panel className="relative flex w-full max-w-xs flex-1 flex-col bg-white pt-5 pb-4">
+                    <Transition.Child
+                      as={Fragment}
+                      enter="ease-in-out duration-300"
+                      enterFrom="opacity-0"
+                      enterTo="opacity-100"
+                      leave="ease-in-out duration-300"
+                      leaveFrom="opacity-100"
+                      leaveTo="opacity-0"
+                    >
+                      <div className="absolute top-0 right-0 -mr-12 pt-2">
+                        <button
+                          type="button"
+                          className="ml-1 flex h-10 w-10 items-center justify-center rounded-full focus:outline-none focus:ring-2 focus:ring-inset focus:ring-white"
+                          onClick={() => setSidebarOpen(false)}
+                        >
+                          <span className="sr-only">Close sidebar</span>
+                          <XMarkIcon
+                            className="h-6 w-6 text-white"
+                            aria-hidden="true"
+                          />
+                        </button>
+                      </div>
+                    </Transition.Child>
+                    <div className="flex flex-shrink-0 items-center px-4">
+                      <img
+                        className="h-6 w-auto"
+                        src={hitsaveLogo}
+                        alt="HitSave"
+                      />
+                    </div>
+                    <div className="mt-5 h-0 flex-1 overflow-y-auto">
+                      <nav className="space-y-1 px-2">
+                        {navigation.map((item) => (
+                          <a
+                            key={item.name}
+                            href={item.href}
+                            className={clsx(
+                              item.current && "bg-gray-100 text-gray-900",
+                              !item.current &&
+                                "text-gray-600 hover:bg-gray-50 hover:text-gray-900",
+                              "group flex items-center px-2 py-2 text-base font-medium rounded-md"
+                            )}
+                          >
+                            <item.icon
+                              className={clsx(
+                                item.current && "text-gray-500",
+                                !item.current &&
+                                  "text-gray-400 group-hover:text-gray-500",
+                                "mr-4 flex-shrink-0 h-6 w-6"
+                              )}
+                              aria-hidden="true"
+                            />
+                            {item.name}
+                            {item.comingSoon && (
+                              <p className="ml-3 block whitespace-nowrap rounded-lg bg-slate-50 py-0.5 px-2 text-xs leading-6 text-slate-400">
+                                Coming Soon
+                              </p>
+                            )}
+                          </a>
+                        ))}
+                      </nav>
+                    </div>
+                  </Dialog.Panel>
+                </Transition.Child>
+                <div className="w-14 flex-shrink-0" aria-hidden="true">
+                  {/* Dummy element to force sidebar to shrink to fit close icon */}
+                </div>
+              </div>
+            </Dialog>
+          </Transition.Root>
+
+          {/* Static sidebar for desktop */}
+          <div className="hidden md:fixed md:inset-y-0 md:flex md:w-64 md:flex-col">
+            {/* Sidebar component, swap this element with another sidebar if you like */}
+            <div className="flex flex-grow flex-col overflow-y-auto border-r border-gray-200 bg-white pt-5">
+              <div className="flex flex-shrink-0 items-center px-4">
+                <Link to="/">
+                  <img className="h-6 w-auto" src={hitsaveLogo} alt="HitSave" />
+                </Link>
+              </div>
+              <div className="mt-5 flex flex-grow flex-col">
+                <nav className="flex-1 space-y-1 px-2 pb-4">
+                  {navigation.map((item) => (
+                    <a
+                      key={item.name}
+                      href={item.href}
+                      className={clsx(
+                        item.current && "bg-gray-100 text-gray-900",
+                        !item.current &&
+                          "text-gray-600 hover:bg-gray-50 hover:text-gray-900",
+                        "group flex items-center px-2 py-2 text-sm font-medium rounded-md"
+                      )}
+                    >
+                      <item.icon
+                        className={clsx(
+                          item.current && "text-gray-500",
+                          !item.current &&
+                            "text-gray-400 group-hover:text-gray-500",
+                          "mr-3 flex-shrink-0 h-6 w-6"
+                        )}
+                        aria-hidden="true"
+                      />
+                      {item.name}
+                      {item.comingSoon && (
+                        <p className="ml-3 block whitespace-nowrap rounded-lg bg-slate-50 py-0.5 px-2 text-xs leading-6 text-slate-400">
+                          Coming Soon
+                        </p>
+                      )}
+                    </a>
+                  ))}
+                </nav>
+              </div>
+            </div>
+          </div>
+          <div className="flex flex-1 flex-col md:pl-64">
+            <div className="sticky top-0 z-10 flex h-16 flex-shrink-0 bg-white shadow">
+              <button
+                type="button"
+                className="border-r border-gray-200 px-4 text-gray-500 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500 md:hidden"
+                onClick={() => setSidebarOpen(true)}
+              >
+                <span className="sr-only">Open sidebar</span>
+                <Bars3BottomLeftIcon className="h-6 w-6" aria-hidden="true" />
+              </button>
+              <div className="flex flex-1 justify-between px-4">
+                <div className="flex flex-1">
+                  <form className="flex w-full md:ml-0" action="#" method="GET">
+                    <label htmlFor="search-field" className="sr-only">
+                      Search
+                    </label>
+                    <div className="relative w-full text-gray-400 focus-within:text-gray-600">
+                      <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center">
+                        <MagnifyingGlassIcon
+                          className="h-5 w-5"
+                          aria-hidden="true"
+                        />
+                      </div>
+                      <input
+                        id="search-field"
+                        className="block h-full w-full border-transparent py-2 pl-8 pr-3 text-gray-900 placeholder-gray-500 focus:border-transparent focus:placeholder-gray-400 focus:outline-none focus:ring-0 sm:text-sm"
+                        placeholder="Search"
+                        type="search"
+                        name="search"
+                      />
+                    </div>
+                  </form>
+                </div>
+                <div className="ml-4 flex items-center md:ml-6">
+                  <button
+                    type="button"
+                    className="rounded-full bg-white p-1 text-gray-400 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
+                  >
+                    <span className="sr-only">View notifications</span>
+                    <BellIcon className="h-6 w-6" aria-hidden="true" />
+                  </button>
+
+                  <ProfileDropdown user={user} />
+                </div>
+              </div>
+            </div>
+
+            <main className="flex-1">{children}</main>
+          </div>
+        </div>
+      </>
+    </>
+  );
+};
