+++ web/src/routes/profile.tsx
@@ -1,0 +1,3 @@
+export default function Profile() {
+  return <div>Welcome user.</div>;
+}
